+++ app/src/main/java/com/example/lastfmuselessapp/model/NavbarItem.kt
@@ -1,6 +1,5 @@
 package com.example.lastfmuselessapp.model
 
 import androidx.compose.ui.graphics.vector.ImageVector
-import com.example.lastfmuselessapp.ui.Screen
 
 data class NavbarItem(val screen: Screen, val icon: ImageVector)
