+++ app/src/main/java/com/example/lastfmuselessapp/model/Screen.kt
@@ -1,0 +1,14 @@
+package com.example.lastfmuselessapp.model
+
+import androidx.annotation.StringRes
+import com.example.lastfmuselessapp.R
+
+sealed class Screen(val route: String, @StringRes val resourceId: Int) {
+    object Home : Screen("home", R.string.home)
+    object Artist : Screen("artist/{artistId}", R.string.artist) {
+        fun getArtistIdArgument(): String = "artistId"
+        fun getRouteForArtistId(artistId: String) = "artist/$artistId"
+    }
+
+    object Library : Screen("library", R.string.library)
+}
